Added the option for moderators to be added to the AUTOYT list

+++ modules/yt-notifications.js
@@ -5,7 +5,7 @@ const res = new (require("rss-parser"))();
 
 module.exports = async (client) => {
     const guild = client.guilds.cache.get(process.env.GUILD_ID);
-    const staffChan = guild.channels.cache.get(process.env.TEST_CHAN);
+    const staffChan = guild.channels.cache.get(process.env.STAFF_CHAN);
     const staffPromoChan = guild.channels.cache.get(process.env.STAFF_PROMO);
     const boostPromoChan = guild.channels.cache.get(process.env.BOOST_PROMO);
 
@@ -18,7 +18,7 @@ module.exports = async (client) => {
 
                 // if the user isn't a booster or staff member, we can remove them from the database
                 const member = guild.members.cache.get(userId);
-                if (!member?.roles?.cache.has(process.env.BOOST_ROLE) && !member?.roles?.cache.has(process.env.STAFF_ROLE)) {
+                if (!member?.roles?.cache.has(process.env.BOOST_ROLE) && !member?.roles?.cache.has(process.env.STAFF_ROLE) && !member?.roles?.cache.has(process.env.MOD_ROLE)) {
                     await ytNotificationSchema.findOneAndRemove({ userId })
                         .catch(err => console.error(`${path.basename(__filename)} There was a problem removing a database entry: `, err));
